requirements.txt
# Cartesia - 2D Voxel Terrain Game
# Core dependencies
pygame>=2.5.0
perlin-noise>=1.12
numpy>=1.24.0
PyYAML>=6.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# mypy>=1.5.0
